[SCons] Configure macOS target version for Python
In some cases, Python can issue a warning about incompatible ABI
versions on macOS when Python is compiled with one target and Cantera is
compiled with another. This change pulls the configuration value from
Python if the full Python package is being built, if it isn't set by an
environment variable.