[SCons/Cython] Use pip to build/install interface
Setuptools versions greater than 60.0.0 deprecate using setup.py install
commands, in preference for PEP 517/518 build backends/frontends. In
this case, we're using pip as the frontend and setuptools as the
backend.

This also uses pip to determine the installation location for the
module. It uses pip's internal API which is subject to change. However,
pip does not expose an external API to get this information, and since
Pip is the one determining where files will be installed, this seems the
most reasonable approach for now.