[SCons/Cython] Change default python_prefix value
When using pip as the installer, setting the default python_prefix equal
to the prefix configuration value is no longer necessary. By removing
this setting, we respect the decision of the Linux distro maintainers as
to where Pip should install packages, since Pip reads this information
from the CPython configuration. On macOS, the system Python is locked at
2.7, so isn't supported anyways. On Windows, this was already the
default.